+++ src/components/pages/CreatePost/CreatePost.tsx
@@ -1,0 +1,5 @@
+import { Presenter } from "./Presenter";
+
+export const CreatePost = () => {
+  return <Presenter />;
+};
