+++ src/components/pages/Page404/Page404.tsx
@@ -1,0 +1,12 @@
+import { useHistory } from "react-router";
+
+import { Presenter } from "./Presenter";
+
+export const Page404 = () => {
+  const history = useHistory();
+  const onClickTop = () => {
+    history.push("/");
+  };
+
+  return <Presenter onClickTop={onClickTop} />;
+};
