+++ src/components/pages/Page404/Presenter.tsx
@@ -1,17 +1,16 @@
 import { VFC } from "react";
-import styled from "styled-components";
-import { useHistory } from "react-router-dom";
 
-import Error404 from "../../images/404.svg";
-import { index as Typography } from "../atom/typography/index";
-import { PrimaryButton } from "../atom/button/PrimaryButton";
+import Error404 from "../../../images/404.svg";
+import { index as Typography } from "../../atom/typography/index";
+import { PrimaryButton } from "../../atom/button/PrimaryButton";
 
-export const Page404: VFC = () => {
-  const history = useHistory();
-  const onClickTop = () => {
-    history.push("/");
-  };
+import { Styled404, StyledFigure } from "./Styles";
 
+type Props = {
+  onClickTop: () => void;
+};
+
+export const Presenter: VFC<Props> = ({ onClickTop }) => {
   return (
     <Styled404>
       <Typography size="1rem">
@@ -29,16 +28,3 @@ export const Page404: VFC = () => {
     </Styled404>
   );
 };
-
-const Styled404 = styled.div`
-  width: 90%;
-  margin: 0 auto;
-  padding: 40px 0;
-  text-align: center;
-`;
-
-const StyledFigure = styled.figure`
-  width: 90%;
-  max-width: 600px;
-  margin: 0 auto;
-`;
