+++ src/components/pages/Page404/Styles.tsx
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const Styled404 = styled.div`
+  width: 90%;
+  margin: 0 auto;
+  padding: 40px 0;
+  text-align: center;
+`;
+
+export const StyledFigure = styled.figure`
+  width: 90%;
+  max-width: 600px;
+  margin: 0 auto;
+`;
