+++ src/components/pages/SinglePostPage/SinglePostPage.tsx
@@ -1,28 +1,15 @@
 import { VFC, useState, useEffect, useRef } from "react";
 import { useParams, useLocation } from "react-router-dom";
 import { useDispatch } from "react-redux";
-import styled from "styled-components";
-import gfm from "remark-gfm";
-import ReactMarkdown from "react-markdown";
-import { Toaster } from "react-hot-toast";
-import { format } from "date-fns";
 
 import firebase from "firebase";
-import { db } from "../../libs/firebase";
+import { db } from "../../../libs/firebase";
 import {
   addLikedPosts,
   removeLikedPosts,
-} from "../../features/users/userSlice";
-import { index as CodeBlock } from "../atom/code/index";
-import { index as Title } from "../atom/title/index";
-import { index as Loading } from "../atom/loading/index";
-import { Sidebar } from "../organisms/Sidebar";
-import { CommentInputArea } from "../organisms/CommentInputArea";
-import { CommentOutputArea } from "../organisms/CommentOutputArea";
-import { COLOR } from "../../Themes/Color";
-import { DEVICE } from "../../Themes/Device";
-import { Page404 } from "./Page404";
-import { useCurrentUser } from "../../hooks/useCurrentUser";
+} from "../../../features/users/userSlice";
+import { useCurrentUser } from "../../../hooks/useCurrentUser";
+import { Presenter } from "./Presenter";
 
 type PostType = {
   postId: string;
@@ -220,142 +207,20 @@ export const SinglePostPage: VFC = () => {
     });
   };
 
-  if (isLoading) {
-    return <Loading width="60" height="60" />;
-  }
-
-  if (post.postId === "") {
-    return <Page404 />;
-  }
-
   return (
-    <StyledSinglePostPage>
-      <StyledTitleWrap>
-        <StyledTitleInner>
-          <Title headline="h1">{post.title}</Title>
-          <StyledTimestamp>
-            {format(post.timestamp, "yyyy-MM-dd")} に公開
-          </StyledTimestamp>
-        </StyledTitleInner>
-      </StyledTitleWrap>
-      <StyledSinglePostPageInner>
-        <StyledMarkdownContainer ref={element}>
-          <StyledReactMarkdown
-            remarkPlugins={[gfm]}
-            children={post.body}
-            components={{ code: CodeBlock }}
-            className="preview"
-          />
-        </StyledMarkdownContainer>
-        <Sidebar
-          author={author}
-          currentUserId={currentUser.uid}
-          postId={postId}
-          location={location.pathname}
-          likedPosts={currentUser.likedPosts}
-          labels={post.labels}
-          onClickLike={onClickLike}
-          countLikes={count}
-          onMouseEnter={onMouseEnter}
-          isShow={isShow}
-        />
-        <StyledCommentArea>
-          <CommentOutputArea postId={postId} comments={comments} />
-          <CommentInputArea
-            uid={currentUser.uid}
-            postId={postId}
-            avatar={currentUser.avatar}
-            username={currentUser.username}
-          />
-        </StyledCommentArea>
-      </StyledSinglePostPageInner>
-      <Toaster position="bottom-right" reverseOrder={false} />
-    </StyledSinglePostPage>
+    <Presenter
+      isLoading={isLoading}
+      post={post}
+      element={element}
+      author={author}
+      currentUser={currentUser}
+      postId={postId}
+      onClickLike={onClickLike}
+      count={count}
+      onMouseEnter={onMouseEnter}
+      isShow={isShow}
+      comments={comments}
+      location={location}
+    />
   );
 };
-
-const StyledSinglePostPage = styled.main`
-  background-color: ${COLOR.BACKGROUND};
-  display: grid;
-  grid-template-columns: 100%;
-  margin: 0 auto;
-  width: 100%;
-  padding: 40px 0;
-
-  @media ${DEVICE.tabletL} {
-    padding: 0 0 40px;
-  }
-`;
-
-const StyledSinglePostPageInner = styled.div`
-  @media ${DEVICE.tabletL} {
-    width: 95%;
-    margin: 0 auto;
-    min-height: 80vh;
-  }
-
-  @media ${DEVICE.laptop} {
-    display: grid;
-    grid-template-columns: 3fr 1fr;
-    gap: 0 20px;
-    width: 100%;
-    max-width: 900px;
-  }
-
-  @media ${DEVICE.laptopL} {
-    max-width: 1200px;
-  }
-`;
-
-const StyledMarkdownContainer = styled.div`
-  padding: 0 14px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 0;
-  }
-`;
-
-const StyledReactMarkdown = styled(ReactMarkdown)`
-  padding: 60px 14px;
-  border-bottom: 1px solid ${COLOR.BACKGROUND};
-  background-color: ${COLOR.WHITE};
-  margin-bottom: 40px;
-  border-radius: 10px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 60px 40px;
-  }
-`;
-
-const StyledTitleWrap = styled.div`
-  background-color: ${COLOR.BACKGROUND};
-`;
-
-const StyledTitleInner = styled.div`
-  padding: 40px 14px 60px;
-
-  @media ${DEVICE.tabletL} {
-    padding: 40px 0;
-    width: 95%;
-    max-width: 900px;
-    margin: 0 auto;
-    text-align: left;
-  }
-
-  @media ${DEVICE.laptopL} {
-    padding: 80px 0;
-    width: 100%;
-    max-width: 1200px;
-  }
-`;
-
-const StyledTimestamp = styled.span`
-  font-size: 12px;
-  color: ${COLOR.GRAY};
-`;
-
-const StyledCommentArea = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-`;
